Refactor jump calculations. Prefer doubles over decimals to more closely match in-game and reduce casts. Improved acquisition of modified Frame Shift Drive information from Frontier API. Enhance loadout / jump calc unit testing.

+++ DataDefinitions/JumpDetail.cs
@@ -5,14 +5,14 @@ namespace EddiDataDefinitions
     public class JumpDetail
     {
         [PublicAPI]
-        public decimal distance { get; private set; }
+        public double distance { get; private set; }
 
         [PublicAPI]
         public int jumps { get; private set; }
 
         public JumpDetail() { }
 
-        public JumpDetail(decimal distance, int jumps)
+        public JumpDetail(double distance, int jumps)
         {
             this.distance = distance;
             this.jumps = jumps;
